feat: add laravel-permission in project

+++ app/Http/Controllers/CategoryController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Constants\PolicyName;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
 use App\Models\Category;
@@ -12,6 +13,8 @@ class CategoryController extends Controller
 {
     public function index(): View
     {
+        $this->authorize(PolicyName::VIEW_ANY, Category::class);
+
         $categories = Category::query()
             ->select(['id', 'name'])
             ->orderBy('name')
@@ -22,28 +25,33 @@ class CategoryController extends Controller
 
     public function create(): View
     {
+        $this->authorize(PolicyName::CREATE, Category::class);
         return view('categories.create');
     }
 
     public function store(StoreCategoryRequest $request): RedirectResponse
     {
+        $this->authorize(PolicyName::CREATE, Category::class);
         Category::query()->create($request->all());
         return redirect()->route('categories.index');
     }
 
     public function show(Category $category): View
     {
+        $this->authorize(PolicyName::VIEW, Category::class);
         return view('categories.show', compact('category'));
     }
 
 
     public function edit(Category $category): View
     {
+        $this->authorize(PolicyName::UPDATE, Category::class);
         return view('categories.create', compact('category'));
     }
 
     public function update(UpdateCategoryRequest $request, Category $category): RedirectResponse
     {
+        $this->authorize(PolicyName::UPDATE, Category::class);
         $category->update([
             'name' => $request->get('name')
         ]);
@@ -53,6 +61,7 @@ class CategoryController extends Controller
 
     public function destroy(Category $category): RedirectResponse
     {
+        $this->authorize(PolicyName::DELETE, Category::class);
         $category->delete();
         return redirect()->route('categories.index');
     }
